requirements.txt
GitPython>=3.1.0
click>=8.0.0
openai>=1.0.0
markdown>=3.3.0
jinja2>=3.0.0
